ready for use

+++ Rotation.h
@@ -1,6 +1,8 @@
 #ifndef __ROTATION_H__
 #define __ROTATION_H__
 
+#include "Matrix4.h"
+
 class Rotation
 {
 public:
@@ -8,7 +10,7 @@ public:
     Matrix4 matrix;
 
     Rotation();
-    Rotation(int rotationId, double angle, double x, double y, double z);
+    Rotation(int rotationId, float angle, float x, float y, float z);
     friend std::ostream &operator<<(std::ostream &os, const Rotation &r);
 };
